requirements.txt
Flask==2.0.3
python-dotenv==1.0.0
requests==2.28.2
gunicorn
ibmcloudant
ibm_cloud_sdk_core